==> 02_ledados.ipynb <==
# Pré processa dados 

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import notas

# %load const.py
CO_ELEVA = 33178860
CO_PARQUE = 33065837
CO_SAO_BENTO = 33062633
CO_SAO_VICENTE = 33063648
CO_SAO_VICENTE = 33063648
CO_ST_INACIO = 33063729
CO_STO_AGO = 33020566
CO_STO_AGO_INST = 33176825
CO_STO_AGOS_NL = 33087679


CO_MUN = 3304557  # Rio de Janeiro
CO_UF_RIO = 33  # RJ

MIN_ALUNOS = 30  # Mínimo de alunos fazendo Enem para considerar na análise
NUM_MELHORES = 60
ANO_ULT = 2019  # último ano a processar

notas_cols = [
    "NU_NOTA_CN",
    "NU_NOTA_CH",
    "NU_NOTA_LC",
    "NU_NOTA_MT",
    "NU_NOTA_REDACAO",
]

# para poder atualizar em um só lugar
from const import ANO_ULT

## Lendo dados do Enem do Rio de Janeiro

from pathlib import Path


def encontra_arq_enem(ano):

    arq = Path(f"dados/microdados_enem{ano}/DADOS/MICRODADOS_ENEM_{ano}.csv")
    if not arq.exists():  # 2017
        arq = Path(
            f"dados/microdados_enem{ano}/Microdados Enem {ano}/DADOS/MICRODADOS_ENEM_{ano}.csv"
        )
    if not arq.exists():  # 2016
        arq = Path(
            f"dados/microdados_enem{ano}/Microdados_enem_{ano}/DADOS/MICRODADOS_ENEM_{ano}.csv"
        )

    if not arq.exists():  # 2012
        arq = Path(f"dados/microdados_enem{ano}/DADOS/DADOS_ENEM_{ano}.csv")
    assert arq.exists(), f"A porra do Inep não padronizou nome de enem para ano {ano}"
    return arq


# rar em 2013
for ano in range(2012, 2020):
    print(encontra_arq_enem(ano))


def read_enem_rio(ano=ANO_ULT):
    arq = encontra_arq_enem(ano)

    print("Lendo dados de:", arq)
    df_enem_rio = pd.concat(
        df_enem.rename(
            columns={
                "COD_ESCOLA": "CO_ESCOLA",
                "COD_UF_ESC": "CO_UF_ESC",
                "NOTA_CN": "NU_NOTA_CN",
                "NOTA_CH": "NU_NOTA_CH",
                "NOTA_LC": "NU_NOTA_LC",
                "NOTA_MT": "NU_NOTA_MT",
                'ST_CONCLUSAO': 'TP_ST_CONCLUSAO',
                'IN_TP_ENSINO': 'TP_ENSINO',
            }
        ).query('CO_UF_ESC == @CO_UF_RIO')
        for df_enem in pd.read_csv(
            arq,
            chunksize=50000,
            sep="," if ano in (2015, 2014, 2012) else ";",
            encoding="iso8859-1",
        )
    )
    df_enem_rio.dropna(subset=["CO_ESCOLA"], inplace=True)
    df_enem_rio["CO_ESCOLA"] = df_enem_rio.CO_ESCOLA.astype(int)
    return df_enem_rio

%reset


df_enem_rio = read_enem_rio()
df_enem_rio.head()

df_enem_rio.shape

df_sv = df_enem_rio.loc[df_enem_rio.CO_ESCOLA == CO_SAO_VICENTE]
notas = list(filter(lambda x: "NOTA" in x, df_sv.columns.to_list()))
df_sv[notas].mean()

## Lendo censo de escolas

Alguns dados que parecem interessantes:

**TP_SITUACAO_FUNCIONAMENTO**	Situação de funcionamento	Num	1	1 - Em Atividade 2 - Paralisada 3 - Extinta (ano do Censo) 4 - Extinta em Anos Anteriores

TP_DEPENDENCIA	Dependência Administrativa	Num	1	1 - Federal 2 - Estadual 3 - Municipal 4 - Privada

TP_CATEGORIA_ESCOLA_PRIVADA	Categoria da escola privada	Num	1	1 - Particular 2 - Comunitária 3 - Confessional 4 - Filantrópica    - Não aplicável para escolas públicas

IN_MANT_ESCOLA_PRIVADA_EMP	Mantenedora da escola privada - Empresa, grupo empresarial do setor privado ou pessoa física

IN_MANT_ESCOLA_PRIVADA_ONG	Mantenedora da escola privada - Organização não governamental (ONG) - internacional ou nacional. Organização da Sociedade Civil de Interesse Público (Oscip)

IN_MANT_ESCOLA_PRIVADA_SIND	Mantenedora da escola privada - Sindicatos de trabalhadores ou patronais, associações e cooperativas

IN_MANT_ESCOLA_PRIVADA_SIST_S	Mantenedora da escola privada - Sistema S (Sesi, Senai, Sesc, Outros)

IN_MANT_ESCOLA_PRIVADA_S_FINS	Mantenedora da escola privada - Instituições sem fins lucrativos

QT_SALAS_EXISTENTES	Número de salas de aula existentes na escola	Num	4

QT_SALAS_UTILIZADAS	Número de salas utilizadas como salas de aula (dentro e fora do prédio)	Num	4

QT_FUNCIONARIOS	Total de funcionários da escola (inclusive profissionais escolares em sala de aula)

**IN_COMUM_MEDIO_MEDIO**	Escola oferece Ensino Médio - Médio Propedêutico em classes comuns do ensino regular (TP_ETAPA_ENSINO igual a 25,26,27,28 ou 29 e IN_ESPECIAL_EXCLUSIVA=0)	Num	1	0 - Não 1 - Sim


df_escolas = pd.read_csv(
    f"dados/microdados_educacao_basica_{ANO_ULT}/microdados_educacao_basica_{ANO_ULT}/DADOS/ESCOLAS.CSV",
    sep="|",
    encoding="latin1",
    low_memory=False,
    index_col="CO_ENTIDADE",
).query(
    "CO_UF == @CO_UF_RIO and IN_COMUM_MEDIO_MEDIO==1 and TP_SITUACAO_FUNCIONAMENTO==1"
)
df_escolas

## Lê dados históricos de turmas 

primeiro_ano = [25, 30, 35]

import glob
import os.path

import patoolib


def dados_turma(ano, try_rar=True):
    dir_censo = [
        i
        for i in glob.glob(f"dados/*{ano}*")
        if ("censo" in i or "educacao_basica" in i) and "zip" not in i
    ]
    assert (
        len(dir_censo) == 1
    ), f"Só pode ter achado um arquivo e achou {len(dir_censo)}"

    arquivo_turmas = glob.glob(f"{dir_censo[0]}/*{ano}*/DADOS/TURMAS.*")
    if not arquivo_turmas:
        arquivo_turmas = glob.glob(f"{dir_censo[0]}/DADOS/TURMAS.*")

    result = [i for i in arquivo_turmas if i.lower().endswith(".csv")]
    if not result:
        result = [i for i in arquivo_turmas if i.lower().endswith(".zip")]
        if try_rar and not result:
            rar = [i for i in arquivo_turmas if i.lower().endswith(".rar")]
            if rar:
                patoolib.extract_archive(rar[0], outdir=os.path.dirname(rar[0]))
                result = dados_turma(ano, False)

    if result:
        result = result[0]
    else:
        result = None
    return result


for i in range(2007, ANO_ULT + 1):
    print(">", dados_turma(i))
# estrutura diretórios bagunçados anets de 2007

col_turmas = pd.DataFrame(
    {
        "id_etapa": [
            "category",
            "FK_COD_ETAPA_ENSINO",
            "TP_ETAPA_ENSINO",
            "TP_ETAPA_ENSINO",
        ],
        "id_escola": [
            "category",
            "PK_COD_ENTIDADE",
            "CO_ENTIDADE",
            "CO_ENTIDADE",
        ],
        "num_matriculas": [
            "uint8",
            "NUM_MATRICULAS",
            "NU_MATRICULAS",
            "QT_MATRICULAS",
        ],
        "id_municipio": [
            "category",
            "FK_COD_MUNICIPIO",
            "CO_MUNICIPIO",
            "CO_MUNICIPIO",
        ],
        "id_uf": [
            "category",
            "FK_COD_ESTADO",
            "CO_UF",
            "CO_UF",
        ],
        #'is_eja':         ['boolean',  'IN_EJA',               'IN_EJA',          'IN_EJA',          ],
    },
    index=[
        "dtype",
        "2007",
        "2015",
        "2018",
    ],
)


def colunas_turmas(columns):
    columns = set(columns)
    match = col_turmas.applymap(columns.__contains__).all(axis=1)
    assert match.sum() == 1, (
        f"Um e somente um registro de nomes de colunas deveria "
        f"casar com todas as colunas dos dados lidos de disco {str(columns)}\n"
        f"\n{match}"
    )
    colunas_map = col_turmas[match].iloc[0].to_dict()
    return colunas_map

def mapa_de_colunas(ano):
    df_teste = pd.read_csv(dados_turma(ano), sep="|", encoding="latin1", nrows=1)
    map_colunas = colunas_turmas(df_teste.columns)

    return {v: k for k, v in map_colunas.items()}


# mapa_de_colunas(2014), mapa_de_colunas(2015)

def le_turma(ano):
    print(f"Lendo ano {ano}")
    map_colunas = mapa_de_colunas(ano)
    df_t = pd.concat(
        (
            df_t  # .rename(columns=map_colunas).query('id_etapa.isin(@primeiro_ano) and id_uf == @CO_MUN_RIO')\
            # .merge(escolas, left_on='id_escola', right_on='CO_ENTIDADE')
            for df_t in pd.read_csv(
                dados_turma(ano),
                sep="|",
                encoding="latin1",
                chunksize=1000,
                # error_bad_lines=False,
                usecols=map_colunas.keys(),
                # nrows=100 #debug
            )
        ),
        sort=True,
    )
    df_t["ano"] = ano
    return df_t


# df_primeiro_ano_turmas = pd.concat(
#     (le_turma(i, df_escolas.CO_ENTIDADE, primeiro_ano) for i in range(2007, 2019)), #2007 e 2008, 2009 falharam linhas diferentes
#     sort=True)
# print(df_primeiro_ano_turmas.shape)
# df_primeiro_ano_turmas.head()
def le_tudo():
    for i in range(2015, ANO_ULT + 1):
        print(f"Turmas ano {i}")
        le_turma(i)


le_tudo()

def le_turmas_padronizadas(filtro, range_=None):
    if not range_:
        range_ = range(2007, ANO_ULT+1)

    return pd.concat(
        [
            (le_turma(i)
            .rename(
                columns=mapa_de_colunas(i)
            )            .astype(col_turmas.loc['dtype'].to_dict()) #precisa do to_dict?
            .query(filtro)
            .assign(ano=i))
            for i in range_
        ],
        sort=True,
    )

df_primeiro_ano_turmas = le_turmas_padronizadas(
    "id_etapa.isin(@primeiro_ano) and id_uf == @CO_UF_RIO"
)
print(df_primeiro_ano_turmas.shape)
df_primeiro_ano_turmas.head()

## Grava tudo em formato binário em disco

# df_turmas.reset_index().to_feather('dados/turmas2018.feather')
df_enem_rio.reset_index().to_feather(f"dados/enem_rio_{ANO_ULT}.feather")

df_escolas.reset_index().to_feather(f"dados/escolas_rio_{ANO_ULT}.feather")

df_primeiro_ano_turmas.reset_index().to_feather("dados/primeiro_ano.feather")

df_turmas_rio = le_turmas_padronizadas("id_uf == @CO_UF_RIO")

df_turmas_rio.reset_index().to_feather("dados/turmas_rio.feather")

## Grava notas de todos os anos do Enem
Enem no formato atual começou mesmo foi em 2009. Ignorarei outros anos.

df_enem_rio = pd.read_feather(f"dados/enem_rio_2019.feather")

import notas

notas.notas_enem(df_enem_rio)

df_notas_todos_enem = pd.concat(
    [
        notas.notas_enem(pd.read_feather(f"dados/enem_rio_{ano}.feather"))
        for ano in range(2009, ANO_ULT + 1)
    ]
)



df = read_enem_rio(2017)

df.NU_ANO

df_enem_tudo= []
for ano in range(2009, ANO_ULT + 1):
    print(f"{ano=}")
    df = read_enem_rio(ano)
    print("... calcula notas")
    dfs.append(notas.notas_enem(df))
print("concatenando")

dfs[0].NU_ANO.value_counts()

%pdb off
pd.concat(
    notas.notas_enem(read_enem_rio(ano))
    for ano in range(
        ANO_ULT,
         2013 - 1, -1
    )  # enem começa pra valer em 2009, mas dados de 2011 vêm na porra de um formato diferente de todos os outros. E 2012 tem colunas também completamente diferentes
).reset_index(drop=True).to_feather("dados/enem_anos.feather")

%pdb on
notas.notas_enem(read_enem_rio(2019))
